# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Aquachile",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/cleanliness": {
            "get": {
                "tags": [
                    "KPI Cleanliness"
                ],
                "summary": "Show all cleanliness kpi",
                "description": "Returns a list of cleanliness kpi",
                "operationId": "16586d3d04359c09586b17924db5856c",
                "responses": {
                    "200": {
                        "description": "Show all cleanliness kpi."
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cleanliness-line/{id}": {
            "get": {
                "tags": [
                    "KPI Cleanliness"
                ],
                "summary": "Show all cleanliness kpi by line",
                "description": "Returns a list of cleanliness kpi by line",
                "operationId": "08ddb87f0eeb5bb4a72655d3dc611d7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Line ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show all cleanliness kpi by line."
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cleanliness-shift/{id}": {
            "get": {
                "tags": [
                    "KPI Cleanliness"
                ],
                "summary": "Show all cleanliness kpi by shift",
                "description": "Returns a list of cleanliness kpi by shift",
                "operationId": "4812c9582e0ab88251dbebfafe846c21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Shift ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show all cleanliness kpi by shift."
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cleanliness/{id}": {
            "post": {
                "tags": [
                    "KPI Cleanliness"
                ],
                "summary": "Create a new cleanliness KPI",
                "description": "Creates a new cleanliness KPI with the received data",
                "operationId": "654ebafefc55a3b60f6ad1100ac39cf0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Line ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Cleanliness KPI object that needs to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "done",
                                    "comment",
                                    "line_name"
                                ],
                                "properties": {
                                    "done": {
                                        "description": "The cleanliness status",
                                        "type": "boolean"
                                    },
                                    "comment": {
                                        "description": "The cleanliness comment",
                                        "type": "boolean"
                                    },
                                    "uuid": {
                                        "description": "The shift uuid",
                                        "type": "uuid"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "done": true,
                                    "comment": "Este es un comentario",
                                    "uuid": "c6b6f8c0-6f7b-11eb-9439-0242ac130002"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}